requirements.txt
streamlit
sentence-transformers
pdfplumber
numpy
faiss-cpu
pillow